Fixed Data class - 0.12
Added common code to parse bools
Removed functions from Data class since it was redundant. Added those
that were necessary to the rest of the program.
Changed the load functions from a text reader to an XDoccument with an
IEnumerable XElement providing to a LINQ statament
Changed the save functions from streamwriters to XDocument and LINQ to
generate and save the XML file